Limit fielddata cache size to 40% heap by default
As we're storing more data in ES, heap usage is seriously increasing.

LPA is at 2GB, PVB at 3GB now.

This is an attempt to reduce the amount of heap that we use to a
reasonably sensible level.

See http://www.elasticsearch.org/blog/using-elasticsearch-and-logstash-to-serve-billions-of-searchable-events-for-customers/ for more detail as to why.