Curator optimize; Make default options 'safe'
Add a cron for 'curator optimize', as this -- in a typical ELK stack --
will save considerable amounts of heap space by reducing the segment count.

See https://gist.github.com/mikepea/511ebb8b723f69cea3a4 for details on this.

This elasticsearch-formula was also assuming it would be used in an
ELK stack (Logstash+Kibana) - deleting indexes after 90 days is not
necessarily a safe thing to do.

As such, adjust the curator options so that they are:

*) Optimise enabled -- pretty safe.
*) Delete is disabled

Also purge the old curator cron, which assumed that only delete would
be run. Curator requires multiple cron entries.